requirements/requirements.txt
beautifulsoup4==4.3.2
bottlenose==0.6.2
Django==1.7.1
django-extensions==1.4.9
django-floppyforms==1.3.0
django-generic-positions==0.1.6
django-haystack==2.3.1
django-hvad==1.0.0
django-libs==1.65.4
django-pure-pagination==0.2.1
django-registration-redux==1.1
django-review==1.9.2
django-user-media==1.1.5
easy-thumbnails==2.2
elasticsearch==1.3.0
factory-boy==2.4.1
gunicorn==19.1.1
lxml==3.4.1
Pillow==2.6.1
psycopg2==2.5.4
python-ptrace==0.8.1
requests==2.5.1
simplejson==3.6.5
six==1.8.0
South==1.0.2
sqlparse==0.1.14
urllib3==1.10
